Add /stats/database/query_types
Logic is shared with summary_db, and now lives in query_types_db.rs.
