+++ src/routes/stats/database/mod.rs
@@ -9,6 +9,7 @@
 // Please see LICENSE file for your rights under this license.
 
 mod over_time_history_db;
+mod query_types_db;
 mod summary_db;
 
-pub use self::{over_time_history_db::*, summary_db::*};
+pub use self::{over_time_history_db::*, query_types_db::*, summary_db::*};
