+++ src/setup.rs
@@ -161,6 +161,7 @@ fn setup(
             stats::over_time_clients,
             stats::database::get_summary_db,
             stats::database::over_time_history_db,
+            stats::database::query_types_db,
             dns::get_whitelist,
             dns::get_blacklist,
             dns::get_regexlist,
